Renamed simple_creature containing folder
