+++ src/phystypes/simple_creature.js
